+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   `--verbose`.
 - Adds `-v` and `--verbose` for printing out the version and path switched to.  
   By default, however, no news is good news.
+- Lists available Node versions when unable to locate the version requested.
 
 ## 1.1.0 (Apr 22, 2015)
 - Detects current version correctly should it contain a revision number